+++ Compiler/action_parser.h
@@ -16,7 +16,7 @@ struct process_memory {
 public:
 	void set(variables_decl* var);
 	void setTags(variables_decl* vars);
-	void setTagValue(std::string tagname, unsigned long long value);
+	void setTagValue(std::string tagname, size_t value);
 	void set(std::string var_name, std::shared_ptr<void> data_ptr);
 	void setRegisters();
 	bool isTag(std::string tagname);
@@ -24,12 +24,12 @@ public:
 	std::shared_ptr<void> getVarPtr(std::string var_name);
 	std::string getVarType(std::string var_name);
 private:
-	std::map<std::string, unsigned long long> unsigned_numbers;
+	std::map<std::string, size_t> unsigned_numbers;
 	std::map<std::string, long long> signed_numbers;
 	std::map<std::string, double> stored_doubles;
 	std::map<std::string, char> stored_chars;
 	std::map<std::string, std::string> stored_strings;
-	std::map<std::string, unsigned long long> stored_tags;
+	std::map<std::string, size_t> stored_tags;
 	std::map<std::string, std::shared_ptr<void>> data_ptrs;
 };
